Add some NaN checks for safety
This is related to the previous bug fix in the previous commit.

If when the scrollbar is not displayed, then e.target.scrollLeft, scrollTop, the scrollHeight and scrollWidth will be 0.
The division of 0 by 0 will result in a NaN that will flow to scope.$$setCenterColumnsData() and create data mis-alignment.
The final result is that the table appears to be empty.